% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_DRG.r
\name{boxplot_DRG}
\alias{boxplot_DRG}
\title{Boxplot for DRG data}
\usage{
boxplot_DRG(data = "medicare")
}
\arguments{
\item{data}{to choose which average payment is to be plotted}
}
\value{
a plot with average payment
}
\description{
Boxplot for DRG data
}
\examples{
boxplot_DRG(data = 'total')

}
