% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_2.R
\name{boxplot_function}
\alias{boxplot_function}
\title{Boxplot Function
This function produces a boxplot}
\usage{
boxplot_function(df, Category)
}
\arguments{
\item{df}{a dataframe}

\item{Category}{a category / column of interest in df}
}
\value{
a boxplot of DRG code vs Category
}
\description{
Boxplot Function
This function produces a boxplot
}
