% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{cal_stats}
\alias{cal_stats}
\title{Calculates statistics over all of the DRG codes for average Medicare payments}
\usage{
cal_stats(stat)
}
\arguments{
\item{stat}{name of the statistics you want, can be 'Mean', 'Median', or 'Sd'}
}
\value{
A dataframe that includes DRG codes and the statistics you want for average Medicare payments
}
\description{
Calculates statistics over all of the DRG codes for average Medicare payments
}
\examples{
cal_stats('Mean')
}
