% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_2.R
\name{Calculate}
\alias{Calculate}
\title{Calculate function
This function produces a table showing DRG Code vs statistical function}
\usage{
Calculate(df, statistic)
}
\arguments{
\item{df}{a dataframe}

\item{statistic}{an already built-in statistical function}
}
\value{
a table of DRG Code vs statistical function
}
\description{
Calculate function
This function produces a table showing DRG Code vs statistical function
}
