% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.R
\name{closest}
\alias{closest}
\title{Auxiliary function}
\usage{
closest(xv, sv)
}
\arguments{
\item{xv}{A numeric vector}

\item{sv}{A number}
}
\description{
Find the closest value to the value sv in the vector xv
}
