% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_boxplot.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{DRG Boxplot}
\usage{
draw_boxplot(df, vary)
}
\arguments{
\item{df}{a dataframe}

\item{vary}{a payment name you want to draw a boxplo}
}
\value{
A boxplot with payments in different DRG code
}
\description{
DRG Boxplot
}
\examples{
data <- load("Lab2/Lab2/data/DRG_data.csv")

draw_boxplot_1(data, "Average Total Payments")

}
