% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{drg.boxplot}
\alias{drg.boxplot}
\title{Boxplot for DRG code}
\usage{
drg.boxplot(y1)
}
\arguments{
\item{y1}{}
}
\value{
plot
}
\description{
Boxplot for DRG code
}
\examples{
 drg.boxplot("Average.Total.Payments")

}
