% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{drgboxplot}
\alias{drgboxplot}
\title{drgboxplot}
\usage{
drgboxplot(payment = "Average.Medicare.Payments")
}
\arguments{
\item{payment}{the type of payment or charge to plot against the respective \code{DRG.Codes}. Legal values include
"Average.Medicare.Payments" "Average.Total.Payments" and "Average.Covered.Charges"}
}
\value{
a plot with Payment (or charge) vs. DRG code
}
\description{
Makes a boxplot of payments or charges vs \code{DRG.Code} from the DRG_data dataset with argument for payment type.
}
\examples{
medicareplot <- drgboxplot("Average.Medicare.Payments")

medicareplot


}
