% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{drgstats}
\alias{drgstats}
\title{drgstats}
\usage{
drgstats(st = "mean")
}
\arguments{
\item{st}{the type of statistic to be reported. Legal values include "mean" "sd" and "median".}
}
\value{
a dataframe containing 3 columns representing the \code{DRG.Code},
the desired statistic, and the number of observations used in the statistic's calculation
}
\description{
Produces a data frame containing statistics for \code{Average.Medicare.Payments} for each \code{DRG.Code}.
}
\examples{
meanstats <- drgstats("mean")

meanstats


}
