% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2.R
\name{fun1}
\alias{fun1}
\title{Lab 2}
\usage{
fun1(df, y)
}
\arguments{
\item{y}{a string for variable y in the dataframe df}

\item{x}{a string for variable x in the dataframe df}
}
\value{
calculations based on what string is given
}
\description{
Lab 2
}
\examples{

fun1(my_data, "Average.Medicare.Payments")

fun3(my_data, "mean")


}
