% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{func1}
\alias{func1}
\title{function1
Make a boxplot of payments by DRG code
Make it an option for your function to do this for either the average Medicare payments,
the average total payment, or the average covered charges}
\usage{
func1()
}
\value{
A boxplot of payments by DRG code
}
\description{
function1
Make a boxplot of payments by DRG code
Make it an option for your function to do this for either the average Medicare payments,
the average total payment, or the average covered charges
}
