% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.R
\name{function_2}
\alias{function_2}
\title{Function_2
This function calculates statistics over all of the DRG codes for average Medicare payments.
Make it an option  to calculate either the mean, median, or standard deviation of the DRG codes.}
\usage{
function_2(x, data)
}
\arguments{
\item{x}{a string from c("mean","median","standard deviation")}

\item{data}{a data frame}
}
\description{
Function_2
This function calculates statistics over all of the DRG codes for average Medicare payments.
Make it an option  to calculate either the mean, median, or standard deviation of the DRG codes.
}
\examples{

mean(data$`Average Medicare Payments`)
median(data$`Average Medicare Payments`)
sd(data$`Average Medicare Payments`)
}
