% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iu.R
\docType{data}
\name{iu}
\alias{iu}
\title{137Cs REM-2 recombination chamber IU characteristic}
\format{
A data frame
\describe{
 \item{Voltage}{numerical - Voltage value}
 \item{Value}{numerical - Current}
}
}
\usage{
data(iu)
}
\description{
The data set contains 137Cs REM-2 recombination chamber IU characteristic measured on 14.09.2022.
}
\examples{
plot(iu$Voltage, iu$Value)
data(iu)
}
\keyword{datasets}
