\name{lab2-package}
\alias{lab2-package}
\alias{lab2}
\docType{package}
\title{
\packageTitle{lab2}
}
\description{
\packageDescription{lab2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab2}
\packageIndices{lab2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lab2}

Maintainer: \packageMaintainer{lab2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
