% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_stats.R
\name{medicare_stats}
\alias{medicare_stats}
\title{Medicare Payment Statistics}
\usage{
medicare_stats(df, x, y, stat)
}
\arguments{
\item{df}{A data frame}

\item{x}{The desired variable to be grouped by, passed as a string}

\item{y}{The variable that is having its statistics calculated, as a string}

\item{stat}{The desired statistic being calculated, passed as a string}
}
\value{
A table of the DRG codes and the desired statistic for each
}
\description{
Calculates the Statistics for DRG codes by Average Medicare Payment
}
\examples{
load("lab_data.Rdata")
medicare_stats(lab2_data,"DRG.Definition","Average.Medicare.Payments","mean")
}
