% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations for bootstrap resampling.}

\item{x}{Vector of data values.}

\item{fun}{Function to apply on the bootstrap samples (default is "mean").}

\item{alpha}{Confidence level (default is 0.05).}

\item{cx}{Expansion factor for text size (default is 1.5).}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\value{
A list containing the confidence interval (ci), applied function (fun), and original data (x).
}
\description{
This function performs bootstrap resampling to estimate a statistic and visualize the results using a histogram.
}
\examples{
myboot2(iter = 10000, x = my_data, fun = "median", alpha = 0.01, cx = 1.2)
}
