% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{myclt}
\alias{myclt}
\title{Title
Generate samples using Central Limit Theorem}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{Number of samples}

\item{iter}{Number of iterations}
}
\value{
Sum of samples
}
\description{
Title
Generate samples using Central Limit Theorem
}
\examples{
myclt(5, 1000)
}
