% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5Task7.R
\name{myhyper}
\alias{myhyper}
\title{Lab5 Task 7 Function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of Iterations.}

\item{N}{Total number of items in the population.}

\item{r}{Total number of sucesses in the population.}

\item{n}{Number of Samples.}
}
\value{
Proportions of successes.
}
\description{
Lab5 Task 7 Function
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
