% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab10Task8.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimation for Normal Distribution Parameters}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{Numeric vector. Sample data.}

\item{mu}{Numeric vector. Grid values for the mean parameter to evaluate likelihood.}

\item{sig}{Numeric vector. Grid values for the standard deviation parameter to evaluate likelihood.}

\item{...}{Additional arguments to be passed to the contour plot function.}
}
\value{
A list containing:
\item{x}{The sample vector.}
\item{coord}{The coordinates of the maximum likelihood estimate in the grid.}
\item{maxl}{The maximum likelihood value.}
\item{muest}{The maximum likelihood estimate for the mean parameter.}
\item{sigest}{The maximum likelihood estimate for the standard deviation parameter.}
}
\description{
This function performs maximum likelihood estimation for the parameters of a normal distribution, given a sample vector.
It plots the likelihood contour and identifies the maximum likelihood estimates.
}
\examples{
mymlnorm(c(1, 2, 3, 4, 5), mu = seq(1, 10, 0.5), sig = seq(0.5, 3, 0.1))
}
