% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Task 7 Lab 6}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{The upper limit of integration (the point at which to stop shading the area under the curve).}

\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution.}
}
\value{
A list containing the mean (`mu`), standard deviation (`sigma`), and the area under the curve up to `a` (`area`).
}
\description{
Task 7 Lab 6
}
\examples{
myncurve(1, 0, 1)
}
