% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{The total number of tickets available.}

\item{gamma}{The risk tolerance level.}

\item{p}{The probability threshold for ticket sales.}
}
\value{
A list containing the optimal number of tickets to be sold under discrete and continuous distribution assumptions.
  The list includes:
  - 'nd': The optimal number of tickets sold under discrete distribution.
  - 'nc': The optimal number of tickets sold under continuous distribution.
  - 'N': The total number of tickets available.
  - 'p': The probability threshold for ticket sales.
  - 'gamma': The risk tolerance level.
}
\description{
Project
}
\examples{
ntickets(400, 0.02, 0.95)
}
