% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pay_boxplot}
\alias{pay_boxplot}
\title{Make a boxplot of payments by DRG code}
\usage{
pay_boxplot(type)
}
\arguments{
\item{type}{type of payments you want, can be 'Average Covered Charges', 'Average Total Payments', or 'Average Medicare Payments'}
}
\value{
A boxplot of payments by DRG code
}
\description{
Make a boxplot of payments by DRG code
}
\examples{
pay_boxplot('Average Covered Charges')
}
