% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2.R
\name{payment_boxplot}
\alias{payment_boxplot}
\title{payment_boxplot}
\usage{
payment_boxplot(
  payment = c("Average.Total.Payments", "Average.Covered.Charges",
    "Average.Medicare.Payments")
)
}
\arguments{
\item{payment}{what payment to make a boxplot (Average.Total.Payments/Average.Covered.Charges/Average.Medicare.Payments)}
}
\value{
a boxplot of payments by DRG code
}
\description{
makes a boxplot of payments by DRG code
}
\examples{
result <- payment_boxplot("Average.Total.Payments")
}
