% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_function.R
\name{recombination_characteristic}
\alias{recombination_characteristic}
\title{Simulate the recombination chamber I(U) characteristic (ionization chamber is a radiation detector operating in an ion recombination regime)}
\usage{
recombination_characteristic(voltage, L = 2.2, noise_level = 1e-05)
}
\arguments{
\item{voltage}{A vector of applied voltages.}

\item{L}{A scalar value representing the constant in the recombination equation (default is 2.2 as REM-2 type recombination chamber parameter).}

\item{noise_level}{A scalar value representing the noise level (default is 0.001).}
}
\value{
An object of class "recombination_characteristic" with attributes "recombination_parameter" and "noise_level".
}
\description{
This function generates a simulated recombination chamber characteristic curve based on the applied voltage.
The result is returned as an object of class "recombination_characteristic".
}
\examples{
# Generate recombination chamber characteristic
characteristic <- recombination_characteristic(seq(100, 200, by = 10))

}
