% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_function.R
\name{stat_function}
\alias{stat_function}
\title{Statistics function}
\usage{
stat_function(df, s)
}
\arguments{
\item{df}{a dataframe}

\item{s}{a calculate method (mean, median, sd)}
}
\value{
a table contains mean or median or sd of DRG codes
}
\description{
Statistics function
}
\examples{
data <- load("Lab2/Lab2/data/DRG_data.csv")

stat_function(data, "mean")

}
