% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2.R
\name{statistics}
\alias{statistics}
\title{Statistics}
\usage{
statistics(method = c("mean", "median", "sd"))
}
\arguments{
\item{method}{what to calculate (mean/median/sd)}
}
\value{
a table of the DRG code and the corresponding result
}
\description{
calculates statistics over all of the DRG codes for average Medicare payments
}
\examples{
result <- statistics("mean")
}
