% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_DRG_Med.r
\name{statistics_DRG_Med}
\alias{statistics_DRG_Med}
\title{Statistics for DRG average Medicare data}
\usage{
statistics_DRG_Med(statistics = "mean")
}
\arguments{
\item{statistics}{to choose which statistic is to be returned}
}
\value{
statistics chosen
}
\description{
Statistics for DRG average Medicare data
}
\examples{
statistics_DRG_Med(statistics = 'median')

}
