% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2.R
\name{statistics_payment}
\alias{statistics_payment}
\title{Statistics for average Medicare payments}
\usage{
statistics_payment(variable_name, data, stat_type)
}
\arguments{
\item{variable_name}{a variable name}

\item{data}{a data frame}

\item{stat_type}{the type of statistics}
}
\value{
Statistics over all of the DRG codes for average Medicare payments.
}
\description{
This function produce a calculates statistics over all of the DRG codes for average Medicare payments.
}
\examples{
statistics_payment("Average Medicare Payments", DRG_data, "mean")
}
