% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Stats}
\usage{
summary_stats(df, stat)
}
\arguments{
\item{df}{a data frame of charges associated with hospital care, including a
column labeled "DRG.Definition" that lists the Diagnosis-Related Group (DRG)
for each charge, and a column labeled "Average.Medicare.Payments"}

\item{stat}{a metric to be calculated for the "Average.Medicare.Payments"
column of the given data frame by Diagnosis-Related Group. Accepts argument
of 'mean', 'median' or 'stdev'.}
}
\value{
a data frame containing one column of Diagnosis-Related Group labels
and one column containing either the mean, median, or standard deviation
of the input variable "Average.Medicare.Payments", to 3 decimal places.
}
\description{
Average, median, or standard deviation of medicare payments by DRG.
}
\examples{
Example data set retrieved from:
https://data.cms.gov/Medicare-Inpatient/Inpatient-Prospective-Payment-System-IPPS-Provider/97k6-zzx3
drg = data(DRG_data.csv)
summary_stats(drg, 'stdev')
}
